Adding "How We Work" starting copy
This language is copied from the [EIS-R (DPA) repo](https://github.com/AlaskaDHSS/EIS-Modernization/blob/master/how-we-work.md). It serves as a good starting point that OCS can use to define the way they intend to work during the Dashboard procurement. The text included has been iterated upon throughout the DPA project, so it reflects lessons learned and the way DPA approaches the work now. 

OCS can use this as their starting point and use as much as is useful and change as much as necessary. @tracymack with DPA could share insights on how it has worked for DPA.