Adding "Code review" copy
This language is copied from the EIS-R (DPA) repo. It serves as a good starting point that OCS can use to define the way they intend to conduct code reviews during the Dashboard procurement. 

OCS can use this as their starting point and use as much as is useful and change as much as necessary. @cnelsonakgov could share insights on how it has worked for the EIS-R (DPA) project.